Update make_cedrcpp.sh; fix some warnings.

+++ cedr/cedr_test_randomized.cpp
@@ -35,7 +35,7 @@ void TestRandomized::init_tracers_vector () {
   };
   Int tracer_idx = 0;
   for (Int perturb = 0; perturb < 6; ++perturb)
-    for (Int ti = 0; ti < sizeof(pts)/sizeof(*pts); ++ti) {
+    for (size_t ti = 0; ti < sizeof(pts)/sizeof(*pts); ++ti) {
       Tracer t;
       t.problem_type = pts[ti];
       const bool shapepreserve = t.problem_type & PT::shapepreserve;
